+++ apps/chat-web/src/components/library/use-library-actions.ts
@@ -1,65 +1,12 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { useNavigate } from '@tanstack/react-router'
-import { createServerFn } from '@tanstack/react-start'
-import { z } from 'zod'
 
-import { graphql } from '../../gql'
 import { useTranslation } from '../../i18n/use-translation-hook'
-import { backendRequest } from '../../server-functions/backend'
 import { toastError, toastSuccess } from '../georgeToaster'
 import { getLibrariesQueryOptions } from './get-libraries'
 import { getLibraryQueryOptions } from './get-library'
 import { deleteLibraryFn } from './server-functions/delete-library'
-
-const updateLibraryParticipants = createServerFn({ method: 'POST' })
-  .inputValidator((data: { libraryId: string; userIds: string[] }) =>
-    z
-      .object({
-        libraryId: z.string(),
-        userIds: z.array(z.string()),
-      })
-      .parse(data),
-  )
-  .handler((ctx) =>
-    backendRequest(
-      graphql(`
-        mutation updateLibraryParticipants($libraryId: String!, $userIds: [String!]!) {
-          updateLibraryParticipants(libraryId: $libraryId, userIds: $userIds) {
-            totalParticipants
-            addedParticipants
-            removedParticipants
-          }
-        }
-      `),
-      {
-        libraryId: ctx.data.libraryId,
-        userIds: ctx.data.userIds,
-      },
-    ),
-  )
-
-const removeLibraryParticipantFn = createServerFn({ method: 'POST' })
-  .inputValidator((data: { libraryId: string; participantId: string }) =>
-    z
-      .object({
-        libraryId: z.string(),
-        participantId: z.string(),
-      })
-      .parse(data),
-  )
-  .handler((ctx) =>
-    backendRequest(
-      graphql(`
-        mutation removeLibraryParticipant($libraryId: String!, $participantId: String!) {
-          removeLibraryParticipant(libraryId: $libraryId, participantId: $participantId)
-        }
-      `),
-      {
-        libraryId: ctx.data.libraryId,
-        participantId: ctx.data.participantId,
-      },
-    ),
-  )
+import { removeLibraryParticipantFn, updateLibraryParticipantsFn } from './server-functions/participants'
 
 export const useLibraryActions = (libraryId: string) => {
   const { t } = useTranslation()
@@ -84,7 +31,7 @@ export const useLibraryActions = (libraryId: string) => {
 
   const { mutate: updateParticipants, isPending: isUpdatingParticipants } = useMutation({
     mutationFn: async ({ userIds }: { userIds: string[] }) => {
-      return await updateLibraryParticipants({ data: { libraryId, userIds } })
+      return await updateLibraryParticipantsFn({ data: { libraryId, userIds } })
     },
     onError: (error) => {
       toastError(t('errors.updateParticipantsFailed', { error: error.message }))
